pyproject.toml
[tool.poetry]
name = "dj-walks-into-bar"
version = "0.1.0"
description = ""
authors = ["Maciej Wrzesniewski <mc.wrzesniewski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

alembic = "1.4.3"
fastapi = "0.61.2"
mysqlclient = "^2.0.1"
SQLAlchemy = "1.3.20"
SQLAlchemy-Utils = "0.36.8"

[tool.poetry.dev-dependencies]
black = "20.8b1"
factory-boy = "3.1.0"
freezegun = "^1.0.0"
ipython = "^7.19.0"
isort = "5.6.4"
mycli = "^1.22.2"
pytest = "^5.2"
pytest-icdiff = "0.5"
pytest-sugar = "0.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
